Configure bulk lookup script via command line arguments and env vars
Configuration via environment variables will help if the script needs to be
executed on a server or as a cloud function. In development, the '.env' file can
be used to manage environment variables. The path to the transactions file can
be set as the first argument to the script or via env var.